requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.2
django-crispy-forms==1.14.0
django-environ==0.8.1
django-phonenumber-field==6.1.0
idna==3.3
PyJWT==2.3.0
pytz==2021.3
requests==2.27.1
secure-smtplib==0.1.1
sqlparse==0.4.2
twilio==7.7.0
urllib3==1.26.8
django-heroku
